refactor aps module
